# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  handler.createFlowRequest:
    properties:
      changeNames:
        items:
          type: string
        type: array
      projectId:
        type: integer
      providerId:
        type: integer
      scheduledAt:
        description: RFC3339
        type: string
    required:
    - changeNames
    - projectId
    - providerId
    - scheduledAt
    type: object
...
  handler.errorResponse:
    properties:
      error:
        type: string
    type: object
  handler.flowItemResponse:
    properties:
      changeName:
        type: string
      finishedAt:
        type: string
      id:
        type: integer
      log:
        type: string
      position:
        type: integer
      startedAt:
        type: string
      status:
        type: string
    type: object
  handler.flowResponse:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/handler.flowItemResponse'
        type: array
      missed:
        type: boolean
      projectId:
        type: integer
      projectName:
        type: string
      providerId:
        type: integer
      scheduledAt:
        type: string
      status:
        type: string
    type: object
...
  handler.reorderFlowRequest:
    properties:
      itemIds:
        items:
          type: integer
        type: array
    required:
    - itemIds
    type: object
...
info:
  contact: {}
  description: REST + SSE API for the OpenSpec Studio backend.
  title: OpenSpec Studio API
  version: "1.0"
paths:
...
  /specflow/flows:
    get:
      description: Lists scheduled flows, optionally filtered to one project.
      parameters:
      - description: Filter to a single registered project
        in: query
        name: projectId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.flowResponse'
            type: array
      summary: List Specflow flows
      tags:
      - specflow
    post:
      consumes:
      - application/json
      description: Only fully-planned changes can be scheduled, and dependency order
        is validated before the flow is saved.
      parameters:
      - description: Flow
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.createFlowRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.flowResponse'
        "400":
          description: incomplete change or unmet dependency
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Schedule a new Specflow flow
      tags:
      - specflow
  /specflow/flows/{id}:
    delete:
      parameters:
      - description: Flow id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      summary: Delete a Specflow flow
      tags:
      - specflow
    get:
      parameters:
      - description: Flow id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.flowResponse'
      summary: Get a Specflow flow
      tags:
      - specflow
  /specflow/flows/{id}/cancel:
    post:
      description: Terminates the in-progress provider process and marks the flow
        cancelled; already-succeeded items are left intact.
      parameters:
      - description: Flow id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "409":
          description: flow is not running
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Cancel a running Specflow flow
      tags:
      - specflow
  /specflow/flows/{id}/reorder:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Flow id
        in: path
        name: id
        required: true
        type: integer
      - description: New item order
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.reorderFlowRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: flow already started
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Reorder a pending flow's items
      tags:
      - specflow
...
swagger: "2.0"